bug: 커뮤니티 상세 이모지 선택 타입 정리

+++ app/src/main/java/com/issueissyu/fe/ui/screens/community/detail/CommunityDetailScreen.kt
@@ -141,7 +141,7 @@ fun CommunityDetailScreenContent(
     onCommunityTakedownClick: () -> Unit = {},
     onEmojiAddClick: () -> Unit = {},
     onEmojiPickerDismiss: () -> Unit = {},
-    onEmojiCandidateClick: (Int) -> Unit = {},
+    onEmojiCandidateClick: (Long) -> Unit = {},
     onEmojiApplyClick: () -> Unit = {},
 ) {
     var showDeclarationDialog by remember { mutableStateOf(false) }
